# docs/api-explorer/IBM_Key_Protect_API-swagger.yaml
swagger: '2.0'
info:
  title: Key Protect
  description: 'IBM Key Protect helps you provision encrypted keys for apps across IBM Bluemix. As you manage the lifecycle of your keys, you can benefit from knowing that your keys are secured by cloud-based hardware security modules (HSMs) that protect against theft of information. You can use the Key Protect API to store, generate, and retrieve your secret key material. Secrets can be any type of data that you would like to protect with the service, for example RSA keys or certificates.'
  version: '1.0'
  contact:
    name: Key Protect
    url: 'https://console.ng.bluemix.net/docs/services/keymgmt/index.html'
  license:
    name: IBM-License
    url: 'http://ibm.com/LICENSE-IBM'
  x-ibm-name: key-protect
schemes:
  - https
host: ibm-key-protect.edge.bluemix.net
basePath: /api/v2
paths:
  /secrets:
    get:
      description: '<p>Retrieves a collection of secrets that can be paged through. </p><p><strong>Note:</strong> <code>GET /secrets</code> will not return a secret in the response body. You can retrieve the secret payload with a subsequent <code>GET /secrets/{id}</code> request.</p>'
      operationId: getSecrets
      parameters:
        - $ref: '#/parameters/AuthParam'
        - $ref: '#/parameters/SpaceParam'
        - $ref: '#/parameters/OrgParam'
        - $ref: '#/parameters/CorrelationId'
        - name: limit
          in: query
          type: integer
          description: The maximum number of secrets to retrieve.
          required: false
        - name: offset
          in: query
          type: integer
          description: The number of secrets to skip.
          required: false
      summary: Retrieve a collection of secret resources
      produces:
        - application/vnd.ibm.collection+json
      responses:
        '200':
          description: The secret collection was successfully retrieved.
          headers:
            Key-Total:
              description: The number of secrets in the space.
              type: integer
          schema:
            $ref: '#/definitions/SecretCollection'
        '401':
          description: |
            Your Bluemix access token is invalid or does not have the necessary permissions to access this resource.
        '403':
          description: |
            The space ID is malformed or invalid.
          schema:
            $ref: '#/definitions/ErrorCollection'
    post:
      description: '<p>Creates a new secret with specified secret material. Key Protect may generate a new secret payload on your behalf using the <code>AlgorithmType</code> and <code>AlgorithmMetadata</code> fields if the secret material is not initially provided. When successful, <code>POST /secrets</code> returns the new secret in the response body, unless the Prefer header is set to <code>return=minimal</code>, in which case the operation will not return the secret in the response.</p>'
      operationId: createSecret
      consumes:
        - application/vnd.ibm.kms.secret+json
      parameters:
        - $ref: '#/parameters/AuthParam'
        - $ref: '#/parameters/SpaceParam'
        - $ref: '#/parameters/OrgParam'
        - $ref: '#/parameters/CorrelationId'
        - $ref: '#/parameters/Prefer'
        - name: secret
          in: body
          description: Secret resource information.
          required: true
          schema:
            $ref: '#/definitions/SecretCollection'
      summary: Create a new secret
      responses:
        '201':
          description: The secret was successfully created.
          schema:
            $ref: '#/definitions/Secret'
        '400':
          description: The secret is missing a required field.
          schema:
            $ref: '#/definitions/ErrorCollection'
        '401':
          description: |
            Your Bluemix access token is invalid or does not have the necessary permissions to access this resource.
        '403':
          description: |
            The space ID is malformed or invalid.
          schema:
            $ref: '#/definitions/ErrorCollection'
    head:
      description: '<p>Returns the same HTTP headers as a GET request without returning the entity-body. This operation returns the number of secrets in your space in a header called <code>Key-Total</code>.'
      operationId: getSecretCollectionMetadata
      parameters:
        - $ref: '#/parameters/AuthParam'
        - $ref: '#/parameters/SpaceParam'
        - $ref: '#/parameters/OrgParam'
        - $ref: '#/parameters/CorrelationId'
      summary: Retrieve metadata for a secret collection without fetching secrets
      responses:
        '200':
          description: The metadata was successfully retrieved.
          headers:
            Key-Total:
              description: The number of secrets in the space.
              type: integer
        '401':
          description: |
            Your Bluemix access token is invalid or does not have the necessary permissions to access this resource.
        '403':
          description: |
            The space ID is malformed or invalid.
  '/secrets/{id}':
    get:
      description: Retrieves a secret resource by specifying the ID of the secret.
      operationId: getSecret
      parameters:
        - $ref: '#/parameters/AuthParam'
        - $ref: '#/parameters/SpaceParam'
        - $ref: '#/parameters/OrgParam'
        - $ref: '#/parameters/CorrelationId'
        - name: id
          in: path
          type: string
          format: uuid
          description: The v4 UUID that uniquely identifies the secret.
          required: true
      summary: Retrieve a secret by ID
      produces:
        - application/vnd.ibm.kms.secret+json
      responses:
        '200':
          description: The secret resource was successfully retrieved.
          schema:
            $ref: '#/definitions/Secret'
        '400':
          description: 'The secret resource could not be retrieved due to a malformed, invalid, or missing ID.'
          schema:
            $ref: '#/definitions/ErrorCollection'
        '401':
          description: |
            Your Bluemix access token is invalid or does not have the necessary permissions to access this resource.
        '403':
          description: |
            The space ID is malformed or invalid.
          schema:
            $ref: '#/definitions/ErrorCollection'
        '404':
          description: The secret resource could not be found. Verify that the secret ID specified is valid.
          schema:
            $ref: '#/definitions/ErrorCollection'
    delete:
      description: Deletes a secret resource by specifying the ID of the secret.
      operationId: deleteSecret
      parameters:
        - $ref: '#/parameters/AuthParam'
        - $ref: '#/parameters/SpaceParam'
        - $ref: '#/parameters/OrgParam'
        - $ref: '#/parameters/CorrelationId'
        - $ref: '#/parameters/Prefer'
        - name: id
          in: path
          type: string
          format: uuid
          description: The v4 UUID that uniquely identifies the secret.
          required: true
      summary: Delete a secret by ID
      responses:
        '200':
          description: The secret resource was successfully deleted.
          schema:
            $ref: '#/definitions/Secret'
        '204':
          description: The secret resource was deleted.
        '400':
          description: 'The secret resource could not be deleted due to a malformed, invalid, or missing ID.'
          schema:
            $ref: '#/definitions/ErrorCollection'
        '401':
          description: |
            Your Bluemix access token is invalid or does not have the necessary permissions to access this resource.
        '403':
          description: |
            The space ID is malformed or invalid.
          schema:
            $ref: '#/definitions/ErrorCollection'
        '404':
          description: The secret resource could not be found.
          schema:
            $ref: '#/definitions/ErrorCollection'
definitions:
  SecretCollection:
    required:
      - metadata
      - resources
    properties:
      metadata:
        $ref: '#/definitions/CollectionMetadata'
      resources:
        type: array
        items:
          $ref: '#/definitions/Secret'
  CollectionMetadata:
    required:
      - collectionType
      - collectionTotal
    properties:
      collectionType:
        type: string
      collectionTotal:
        type: integer
        format: int64
  Secret:
    required:
      - type
      - name
    properties:
      type:
        description: 'Specifies the MIME type that represents the secret resource. Currently, only the default is supported.'
        type: string
        default: application/vnd.ibm.kms.secret+json
      id:
        description: The v4 UUID used to uniquely identify the resource, as specified by RFC 4122.
        type: string
        format: uuid
        readOnly: true
      name:
        description: 'A human-readable, non-unique alias to assign to your secret.'
        type: string
        pattern: '[a-zA-Z0-9 ]{2,240}'
        minLength: 2
        maxLength: 240
      creationDate:
        type: string
        format: date-time
        readOnly: true
        example: '2010-01-12T05:23:19+0000'
        description: The date the secret material was created.
      expirationDate:
        type: string
        format: date-time
        example: '2010-01-12T05:23:19+0000'
        description: The date the secret material expires.
      createdBy:
        description: The unique identifier for the resource that created the secret.
        type: string
        readOnly: true
      algorithmType:
        description: 'The algorithm type to be used when generating the secret. Currently, AES is supported.'
        type: string
        enum:
          - AES
        default: AES
      algorithmMetadata:
        type: object
        description: 'The algorithm type to be used when generating the secret. Currently, AES is supported.'
      payload:
        description: Provide the secret material or else leave blank so that Key Protect may generate one for you based on the specified algorithmType and algorithmMetadata fields.
        type: string
        format: binary
      description:
        description: A text field used to provide a more detailed description of the secret. The maximum length is 240 characters.
        type: string
        minLength: 2
        maxLength: 240
      state:
        type: integer
        description: 'The key state based on NIST SP 800-57. States are integers and correspond to the Pre-activation = 0, Active = 1, Deactivated = 3, and Destroyed = 5 values.'
        readOnly: true
        enum:
          - Pre-activation
          - Activation
          - Deactivated
          - Destroyed
      nonactiveStateReason:
        description: A code indicating the reason the secret is not in the activation state.
        type: integer
        readOnly: true
      deleted:
        type: boolean
        readOnly: true
        description: A boolean value that determines whether the key has been deleted.
      activationDate:
        description: 'When an activationDate is included, Key Protect creates a secret in the Pre-activation state before transitioning to the Activation state on the specified date. If no date is specified, the secret is activated immediately. The date is specified by RFC 3339.'
        type: string
        format: date-time
        example: '2010-01-12T05:23:19+0000'
      deletionDate:
        description: The date the secret material was destroyed. The date is specified by RFC 3339.
        type: string
        format: date-time
        readOnly: true
        example: '2010-01-12T05:23:19+0000'
      lastUpdateDate:
        description: The date is specified by RFC 3339 and updated when any part of the secret metadata is modified.
        type: string
        format: date-time
        readOnly: true
        example: '2010-01-12T05:23:19+0000'
      deletedBy:
        type: string
        readOnly: true
        description: The unique identifier for the resource that deleted the secret.
      tags:
        description: 'Up to 30 tags can be created. Tags can be between 2-30 characters, including spaces. Special characters not permitted include the angled bracket, comma, colon, ampersand, and vertical pipe character (|).'
        type: array
        items:
          type: string
  Error:
    properties:
      errorMsg:
        type: string
  ErrorCollection:
    required:
      - metadata
      - resources
    properties:
      metadata:
        $ref: '#/definitions/CollectionMetadata'
      resources:
        type: array
        items:
          $ref: '#/definitions/Error'
parameters:
  AuthParam:
    name: Authorization
    in: header
    type: string
    format: byte
    description: Your Bluemix access token.
    required: true
  OrgParam:
    name: Bluemix-Org
    in: header
    type: string
    description: The Bluemix organization GUID.
    required: true
  SpaceParam:
    name: Bluemix-Space
    in: header
    type: string
    description: The Bluemix space GUID.
    required: true
  CorrelationId:
    name: Correlation-Id
    in: header
    type: string
    format: uuid
    description: The v4 UUID used to correlate and track transactions.
    required: false
  Prefer:
    name: Prefer
    in: header
    description: 'Alters server behavior for POST or DELETE operations. A header with `return=minimal` causes the service to return only the secret identifier, or metadata. A header containing `return=representation` returns the secret payload and metadata in the entity-body. <br><br> **Note:** During POST operations, Key Protect may not immediately return the secret payload due to key generation time. To retrieve the secret payload, you can perform a subsequent `GET /secrets/{id}` request.'
    required: false
    type: string
x-ibm-configuration:
  extensions:
    aes-category: 1.0.0
    aes-metadata: 1.0.0
x-aes-category:
  bluemixList:
    - Security
  technologyList:
    - bluemix
  namespaceList:
    - apiexplorer
    - bluemix
x-aes-metadata:
  summary: IBM Key Protect helps you provision encrypted keys for apps across IBM Bluemix
  apiStatus: Production
  serviceStatus: STAGE_LIVE
  productImage: ''
  linktoProduct: 'https://console.ng.bluemix.net/docs/services/keymgmt/index.html'
  uniqueShortName: IBM_Key_Protect_API
  companyName: IBM
  providerEmail: jhroyal@us.ibm.com
  marketPlaceList:
    - Security
  languagelist:
    - JavaScript + jQuery
    - Python + Requests
    - Shell + Curl
  whoCanEdit:
    - diazjf@us.ibm.com
    - erlander.lo@us.ibm.com
    - kmaeda@us.ibm.com
    - zac.nixon@ibm.com
    - brienjo@us.ibm.com
    - tasalasc@us.ibm.com
    - seguras@us.ibm.com
    - edmund.troche@us.ibm.com
    - crystal.barragan@ibm.com